% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.all.R
\name{install.all}
\alias{install.all}
\title{Install All R Packages within the 'basicPackage'}
\usage{
install.all()
}
\description{
The function includes the following R packages families be installed in the latest version:
c("install.Bioconductor()", "install.data.analysis()", 
"install.gg.series()", "install.github.PhDMeiwp()", "install.graphic()", 
"install.isotope()", "install.others()", 
"install.writedown()", "install.yyplot()")
}
\examples{

library(basicPackages)
basicPackages::install.all()
}
