% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.basicANCOVA.R
\name{install.basicANCOVA}
\alias{install.basicANCOVA}
\title{Install R Package of 'basicANCOVA' from Github}
\usage{
install.basicANCOVA()
}
\description{
The function includes the following R packages be installed in the latest version:
c("devtools", "basicANCOVA")
}
\examples{

library(basicPackages)
basicPackages::install.basicANCOVA()

library(basicANCOVA)
data("isotope",package = "basicANCOVA")
View(isotope)
#rename
groups<-isotope$area
x<-isotope$d13C
y<-isotope$d15N
data<-isotope
ANCOVAplot(x,y,groups,data)
}
