% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.data.analysis.R
\name{install.data.analysis}
\alias{install.data.analysis}
\title{Install R Packages for Data Analysis}
\usage{
install.data.analysis()
}
\description{
The function includes the following R packages be installed in the latest version:
c("base2grob","basicTrendline","bayesm",
"car","caret","colorspace",
"devtools","dplyr","dunn.test",
"fANCOVA","forcats",
"HH",
"labeling","labelled","laercio","latticExtra","lme4","lmtest",
"NSM3",
"parallel", "permute","plot3D","plot3Drgl","plyr","polyclip","polynom","psych",
"R.methodsS3","randomForest","ranger","Rcpp","reshape","reshape2","rmda","robustbase","roxygen2",
"scholar",
"tidyr","tweenr",
"vegan",
"waveslim","whisker","wordcloud",
"xml2","xtable")
}
\examples{

library(basicPackages)
basicPackages::install.data.analysis()
}
