% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.gg.series.R
\name{install.gg.series}
\alias{install.gg.series}
\title{Install R Packages of 'ggplot2' and its extensions}
\usage{
install.gg.series()
}
\description{
The function includes the following R packages be installed in the latest version:
c("ggforce","ggimage","ggplot2",
"ggpubr","ggrepel","ggsci","ggsignif","git2r",
"ggtree", "gridExtra","gridGraphics", "growthcurver")
}
\examples{

library(basicPackages)
basicPackages::install.gg.series()
}
