% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.github.PhDMeiwp.R
\name{install.github.PhDMeiwp}
\alias{install.github.PhDMeiwp}
\title{Install R Packages contained in the repositories of Github user 'PhDMeiwp'}
\usage{
install.github.PhDMeiwp()
}
\description{
The function includes the following R packages be installed in the latest version:
c("devtools","basicANCOVA", "basicTrendline", "circos.JCR", "Steel.Dwass.test", "prettyB", "R11")
}
\examples{

library(basicPackages)
basicPackages::install.github.PhDMeiwp()
}
