% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.graphic.R
\name{install.graphic}
\alias{install.graphic}
\title{Install R Packages for Graphic excluding gg.series}
\usage{
install.graphic()
}
\description{
The function includes the following R packages be installed in the latest version:
c("qrcode","tidyr","tweenr","wordcloud",
"htmlTable","htmltools","htmlwidgets","httpcode")
}
\examples{

library(basicPackages)
basicPackages::install.graphic()
}
