% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.isotope.R
\name{install.isotope}
\alias{install.isotope}
\title{Install R Packages for Stable Isotope Analysis}
\usage{
install.isotope()
}
\description{
The function includes the following R packages be installed in the latest version:
c("IsotopeR", "IsoriX", "MixSIAR", 
"siar", "SIBER", "simmr", 
"tRophicPosition")
}
\examples{

library(basicPackages)
basicPackages::install.isotope()
}
