% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.others.R
\name{install.others}
\alias{install.others}
\title{Install some other R Packages}
\usage{
install.others()
}
\description{
The function includes the following R packages be installed in the latest version:
c("rvcheck")
}
\examples{

library(basicPackages)
basicPackages::install.others()
}
