% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.Steel.Dwass.test.R
\name{install.Steel.Dwass.test}
\alias{install.Steel.Dwass.test}
\title{Install R Package of 'Steel.Dwass.test' from Github}
\usage{
install.Steel.Dwass.test()
}
\description{
The function includes the following R packages be installed in the latest version:
c("devtools", "Steel.Dwass.test")
}
\examples{

library(basicPackages)
basicPackages::install.Steel.Dwass.test()

library(Steel.Dwass.test)
x <- c(6.9, 7.5, 8.5, 8.4, 8.1, 8.7, 8.9, 8.2, 7.8, 7.3, 6.8,9.6, 9.4, 9.5, 8.5, 9.4, 9.9, 8.7, 8.1, 7.8, 8.8,5.7, 6.4, 6.8, 7.8, 7.6, 7.0, 7.7, 7.5, 6.8, 5.9,7.6, 8.7, 8.5, 8.5, 9.0, 9.2, 9.3, 8.0, 7.2, 7.9, 7.8)
group <- rep(1:4, c(11, 10, 10, 11))
Steel.Dwass(x,group)
}
