% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.writedown.R
\name{install.writedown}
\alias{install.writedown}
\title{Install R Packages for Writedown Series}
\usage{
install.writedown()
}
\description{
The function includes the following R packages be installed in the latest version:
c("knitr","blogdown","bookdown","markdown","rmarkdown")
}
\examples{

library(basicPackages)
basicPackages::install.writedown()
}
