% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.yyplot.R
\name{install.yyplot}
\alias{install.yyplot}
\title{Install R Packages of 'yyplot' from Github}
\usage{
install.yyplot()
}
\description{
The function includes the following R packages be installed in the latest version:
c("devtools","digest", "crul", "xml2", "stringi", "yyplot")
}
\examples{

library(basicPackages)
basicPackages::install.yyplot()

library(yyplot)
term <- c('"H7N9"', '"H5N1"', '"RSV"')
pm <- pubmed_trend(term, year=2001:2014)
plot(pm)

pubmed_trend("Yu Guangchuang[Full Author Name]", 2010:2016)

#Link: https://guangchuangyu.github.io/cn/2017/11/pubmed-trend/
}
