% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renviron.R
\name{Renviron}
\alias{Renviron}
\title{Creat '.Renviron' file}
\usage{
Renviron()
}
\description{
Or type file.edit('~/.Renviron') in R.
}
\examples{

# Not run
#-------------------
## Type the following codes in your '.Renviron' without '#'.
R_LIBS_USER="~/R/UserPackages"
#--------------------
# End (Not run)

}
